requirements.txt
langchain-google-genai
langchain-openai
streamlit
python-dotenv
langchain-core
langserve
fastapi
uvicorn
sse_starlette
langchain-community
bs4
pypdf
langchain
chromadb
faiss-cpu
sentence-transformers
huggingface-hub
langchain-classic